events, map

+++ backend/src/main/java/com/inspiral/inspiralbackend/models/Training.java
@@ -34,10 +34,10 @@ public class Training {
     private Date date;
 
     @Column(name = "lat")
-    private Integer lat;
+    private Double lat;
 
     @Column(name = "lon")
-    private Integer lon;
+    private Double lon;
 
 
 
@@ -78,11 +78,11 @@ public class Training {
         return date;
     }
 
-    public Integer getLat() {
+    public Double getLat() {
         return lat;
     }
 
-    public Integer getLon() {
+    public Double getLon() {
         return lon;
     }
 
@@ -114,11 +114,11 @@ public class Training {
         this.date = date;
     }
 
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
         this.lat = lat;
     }
 
-    public void setLon(Integer lon) {
+    public void setLon(Double lon) {
         this.lon = lon;
     }
 }
